requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
httpx==0.26.0
resend==0.7.0
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
